+++ docs/release_notes/changelog.rst
@@ -5,6 +5,9 @@
 Changelog
 =========
 
+UNRELEASED
+
+- added test for :meth:`staircase.Stairs.logical_xor` with two Stairs arguments
 
 v2.0.0 2021-08-25